+++ common/src/main/java/com/izofar/takesapillage/common/world/feature/MobFeature.java
@@ -1,5 +1,7 @@
 package com.izofar.takesapillage.common.world.feature;
 
+import com.google.common.collect.ImmutableList;
+import com.izofar.takesapillage.common.util.MobLists;
 import com.izofar.takesapillage.common.versions.VersionedEntitySpawnReason;
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.entity.EntityType;
@@ -10,27 +12,27 @@ import net.minecraft.world.level.levelgen.feature.configurations.NoneFeatureConf
 import java.util.function.Supplier;
 
 //? if >=1.21.5 {
-import net.minecraft.util.random.Weighted;
+/*import net.minecraft.util.random.Weighted;
 import net.minecraft.util.random.WeightedList;
-//?} else {
-/*import net.minecraft.util.random.WeightedRandomList;
+*///?} else {
+import net.minecraft.util.random.WeightedRandomList;
 import net.minecraft.util.random.WeightedEntry;
-*///?}
+//?}
 
 public class MobFeature<T extends Mob> extends Feature<NoneFeatureConfiguration>
 {
 	//? if >=1.21.5 {
-	private final Supplier<WeightedList<Weighted<EntityType<? extends T>>>> entityTypes;
-	//?} else {
-	/*private final Supplier<WeightedRandomList<WeightedEntry.Wrapper<EntityType<? extends T>>>> entityTypes;
-	*///?}
+	/*private final Supplier<WeightedList<EntityType<? extends T>>> entityTypes;
+	*///?} else {
+	private final Supplier<WeightedRandomList<WeightedEntry.Wrapper<EntityType<? extends T>>>> entityTypes;
+	//?}
 
 	public MobFeature(
 		//? if >=1.21.5 {
-		Supplier<WeightedList<Weighted<EntityType<? extends T>>>> entityTypes
-		//?} else {
-		/*Supplier<WeightedRandomList<WeightedEntry.Wrapper<EntityType<? extends T>>>> entityTypes
-		*///?}
+		/*Supplier<WeightedList<EntityType<? extends T>>> entityTypes
+		*///?} else {
+		Supplier<WeightedRandomList<WeightedEntry.Wrapper<EntityType<? extends T>>>> entityTypes
+		//?}
 	) {
 		super(NoneFeatureConfiguration.CODEC);
 		this.entityTypes = entityTypes;
@@ -38,24 +40,17 @@ public class MobFeature<T extends Mob> extends Feature<NoneFeatureConfiguration>
 
 	public MobFeature(EntityType<? extends T> entityType) {
 		super(NoneFeatureConfiguration.CODEC);
-		//? if >=1.21.5 {
-		this.entityTypes = () -> WeightedList.of(new Weighted<EntityType<? extends T>>(entityType, 1));
-		//?} else {
-		/*this.entityTypes = () -> WeightedRandomList.create(WeightedEntry.<EntityType<? extends T>>wrap(entityType, 1));
-		*///?}
+		this.entityTypes = () -> MobLists.createWeightedList(ImmutableList.of(MobLists.createWeightedEntry(entityType, 1)));
 	}
 
 	@Override
 	public boolean place(FeaturePlaceContext<NoneFeatureConfiguration> context) {
 		BlockPos position = context.origin().below();
-		var entry = this.entityTypes.get().getRandom(context.random()).get();
-
 		//? if >=1.21.5 {
-		EntityType<? extends Mob> entityType = entry.value();
-		//?} else {
-		/*EntityType<? extends Mob> entityType = entry.value();
-		*///?}
-
+		/*var entityType = this.entityTypes.get().getRandom(context.random()).get();
+		*///?} else {
+		var entityType = this.entityTypes.get().getRandom(context.random()).get().data();
+		//?}
 		var entity = entityType.create(context.level().getLevel()/*? >=1.21.3 {*/, VersionedEntitySpawnReason.STRUCTURE/*?}*/);
 
 		if (entity == null) {
@@ -63,10 +58,10 @@ public class MobFeature<T extends Mob> extends Feature<NoneFeatureConfiguration>
 		}
 
 		//? >=1.21.5 {
-		entity.snapTo(position.getX() + 0.5D, position.getY(), position.getZ() + 0.5D, 0.0F, 0.0F);
-		//?} else {
-		/*entity.moveTo(position.getX() + 0.5D, position.getY(), position.getZ() + 0.5D, 0.0F, 0.0F);
-		*///?}
+		/*entity.snapTo(position.getX() + 0.5D, position.getY(), position.getZ() + 0.5D, 0.0F, 0.0F);
+		*///?} else {
+		entity.moveTo(position.getX() + 0.5D, position.getY(), position.getZ() + 0.5D, 0.0F, 0.0F);
+		//?}
 		//? if >=1.21.1 {
 		entity.finalizeSpawn(context.level(), context.level().getCurrentDifficultyAt(position), VersionedEntitySpawnReason.STRUCTURE, null);
 		//?} else {