+++ fabric/src/main/java/com/izofar/takesapillage/fabric/mixin/StructureTemplateEntitySpawnMixin.java
@@ -37,7 +37,7 @@ public class StructureTemplateEntitySpawnMixin
 			),
             cancellable = true
 	)
-    private static void friendsandfoes$onEntitySpawn(
+    private static void takesapillage$onEntitySpawn(
 		Rotation rotation,
 		Mirror mirror,
 		Vec3 vec3,